_curso/17_admin_catalogo_de_videos/python/requirements.txt
pymongo==4.1.1
requests==2.27.1
redis==4.3.4
djangorestframework==3.14.0